// src/app/core/model/species.ts
export interface ISpecies {
  name: string;
  classification: string;
  designation: string;
  average_height: string;
  skin_colors: string;
  hair_colors: string;
  eye_colors: string;
  average_lifespan: string;
  homeworld: number;
  language: string;
  people: number[];
  films: number[];
  created: string;
  edited: string;
  id: number;
}